Update and rename ast_sheet_sim.ts to src/scripts/sims/ast_sheet_sim.ts
